pyproject.toml
[tool.poetry]
name = "radas"
version = "0.1.0"
description = "Plasma radiation calculated using OpenADAS"
authors = ["Commonwealth Fusion Systems"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
numpy = "^1.24.2"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
Pint = "^0.20.1"
pyyaml = "^6.0"
xarray = "^2023.3.0"
pint-xarray = "^0.3"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
coverage = "^7.2.2"
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
